server/requirements.txt
flask==1.1.2
Flask-Admin==1.5.7
flask-cors==3.0.10
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
python-json-logger==2.0.2
psycopg2-binary==2.8.6
PyJWT==1.7.1
pandas==1.3.4
